Commented out all the states, and debugging each state and behavior separately.
The goal keeper currently, waits for the game to begin, then changes
state to tend goal -> intercept goal and the on fetching the ball
returns back to his home region and then waits till the game is on which
is basically every other player is back to their position. After which
he changes state to goal kick.